requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-ckeditor5==0.0.4
django-crispy-forms==2.3
django-grappelli==4.0.1
django-js-asset==2.2.0
django-paypal==2.1
django-shortuuidfield==0.1.3
django-taggit==6.0.0
idna==3.8
pillow==10.4.0
pytz==2024.1
requests==2.32.3
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
